collector/requirements.txt
brotlicffi==1.2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
packaging==25.0
peewee==3.18.2
pika==1.3.1
pip==26.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.23
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.10.2
urllib3==2.6.3
wheel==0.46.3
